Trying a different precision mapping

+++ src/MEDS_EIC_AR/model/model.py
@@ -139,9 +139,9 @@ class Model(torch.nn.Module):
     PRECISION_TO_MODEL_WEIGHTS_DTYPE: ClassVar[dict[str, torch.dtype]] = {
         "32-true": torch.float32,
         "16-true": torch.float16,
-        "16-mixed": torch.float32,
+        "16-mixed": torch.float16,
         "bf16-true": torch.bfloat16,
-        "bf16-mixed": torch.float32,
+        "bf16-mixed": torch.float16,
         "transformer-engine": torch.bfloat16,
     }
